cognitive_speech/requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
azure-ai-formrecognizer      
azure-cognitiveservices-speech
azure-common                  
azure-core                    
azure-functions               
azure-identity                
azure-search                   
azure-search-documents         
azure-storage-blob